Netflix Clone using HTML, CSS & JavaScript
🎥 **Netflix Clone Overview:**

Immerse yourself in the world of entertainment with my Netflix clone, meticulously crafted using HTML, CSS, and JavaScript. The project features a transparent and responsive header that dynamically changes its background as you scroll. The banner section showcases trending movies with captivating visuals and details. Each movie category comes to life with a beautifully designed section, and YouTube trailer integration adds an interactive element to the experience. With a seamless and responsive design, this Netflix clone promises an engaging streaming adventure. 🍿✨